% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkson.logit.R
\name{berkson.logit}
\alias{berkson.logit}
\title{Berkson's logit chi square}
\usage{
berkson.logit(formula, data)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data}
}
\description{
Berkson's logit chi square
}
\examples{
berkson.logit()
}
\keyword{Berkson}
\keyword{logit}

